refactor(mcp-memory): split server.py into handlers/ package (#360)
server.py shrinks from 3466 → 295 lines. Handlers, tool schemas,
embedding pipeline, and Supabase client move to dedicated modules.

Layout:
  mcp-memory/
  ├── server.py            # entry, dispatch, _compute_write_embeddings, main()
  ├── client.py            # _get_client + _audit_log
  ├── embeddings.py        # Voyage AI + _model_slot/_canonical_embed_text
  ├── tools_schema.py      # tool_definitions() — 21 Tool entries
  └── handlers/
      ├── memory.py        # store/recall/get/list/delete/restore/graph + helpers
      ├── goal.py          # 4 goal handlers + GOAL_FIELDS + _format_goal
      ├── outcome.py       # outcome_record/update/list + memory_calibration_summary
      ├── credential.py    # credential_list/add/check_expiry
      ├── events.py        # events_list/mark_processed
      └── decision.py      # _looks_like_uuid + _resolve_memory_refs + record_decision

Test suite passes unmodified (936/936). Pure code-move except: dropped
the dead duplicate sync defs of _cosine_sim/_upsert_known_unknown
that were shadowed at runtime by their async siblings (1508/1520 →
2402/2418); preserved behavior because Python resolves to the second
def at lookup time.

Tool names + JSON schemas unchanged — redrobot consumers unaffected.

Closes #360
