+++ mcp-memory/handlers/__init__.py
@@ -1,0 +1,5 @@
+"""Handler package — one module per tool family (#360 split).
+
+Re-exported into the top-level `server` namespace for backwards
+compatibility with tests and external callers.
+"""
