+++ mcp-memory/handlers/decision.py
@@ -1,0 +1,174 @@
+"""Decision-trace handler — record_decision + name→UUID resolver helpers.
+
+Part of #360 server.py split. Calls utilities (`_get_client`,
+`_audit_log`, embedding helpers, `EMBEDDING_MODEL_*`, ...) via
+the `server` module so test monkeypatches on those names
+propagate at call time.
+"""
+
+from __future__ import annotations
+
+import uuid
+
+from datetime import datetime, timezone  # noqa: F401
+
+from mcp.types import TextContent  # noqa: F401
+
+import server  # noqa: F401  — late-bound for monkeypatch propagation
+
+
+def _looks_like_uuid(s: str) -> bool:
+    """True if s is a canonical UUID string (case-insensitive, hyphenated)."""
+    try:
+        uuid.UUID(s)
+        return True
+    except (ValueError, AttributeError, TypeError):
+        return False
+
+
+def _resolve_memory_refs(client, refs: list, project: str | None) -> tuple[list[str], list[str]]:
+    """Normalize mixed memory names and UUIDs into canonical UUIDs (#325).
+
+    Policy:
+    - Inputs already shaped like a UUID pass through, canonicalized via
+      uuid.UUID() (lower-cased, hyphenated).
+    - Non-UUID strings are treated as memory names and resolved against
+      the memories table, preferring the most recently updated live row
+      (same join heuristic as scripts/backfill-outcome-memories.py).
+    - When ``project`` is provided, the name lookup is scoped to that
+      project. Otherwise the most-recent-live match across all projects
+      wins.
+    - Unresolvable names are returned separately so the handler can
+      surface them without breaking the write.
+
+    Returns (resolved_uuids, unresolved_names). Insertion order is
+    preserved; resolved UUIDs are de-duplicated.
+    """
+    resolved: list[str] = []
+    unresolved: list[str] = []
+    seen: set[str] = set()
+    for ref in refs or []:
+        if not isinstance(ref, str):
+            continue
+        ref_s = ref.strip()
+        if not ref_s:
+            continue
+        if _looks_like_uuid(ref_s):
+            canonical = str(uuid.UUID(ref_s))
+            if canonical not in seen:
+                resolved.append(canonical)
+                seen.add(canonical)
+            continue
+        try:
+            q = client.table("memories").select("id").eq("name", ref_s).is_("deleted_at", "null")
+            if project is not None:
+                q = q.eq("project", project)
+            rows = q.order("updated_at", desc=True).limit(1).execute()
+        except Exception:
+            unresolved.append(ref_s)
+            continue
+        data = getattr(rows, "data", None)
+        if not isinstance(data, list) or not data:
+            unresolved.append(ref_s)
+            continue
+        uid = data[0].get("id") if isinstance(data[0], dict) else None
+        if not uid or not _looks_like_uuid(uid):
+            unresolved.append(ref_s)
+            continue
+        canonical = str(uuid.UUID(uid))
+        if canonical not in seen:
+            resolved.append(canonical)
+            seen.add(canonical)
+    return resolved, unresolved
+
+
+async def _handle_record_decision(args: dict) -> list[TextContent]:
+    """Insert a 'decision_made' episode with structured payload (#252, #325).
+
+    The episode is the agent's reasoning trace: what was decided, why,
+    which memories/outcomes informed it, predicted confidence, and
+    reversibility. /reflect reads these back via the episodes table to
+    analyze whether the basis was sound when outcomes come in.
+
+    ``memories_used`` accepts either UUIDs or memory names — names are
+    resolved server-side so the payload always stores canonical UUIDs,
+    keeping the Pillar 3 FK (``task_outcomes.memory_id``) joinable
+    forward. Unresolved names are surfaced in the response text and
+    preserved on ``payload.memories_used_unresolved`` for audit.
+    """
+    decision = (args.get("decision") or "").strip()
+    rationale = (args.get("rationale") or "").strip()
+    reversibility = args.get("reversibility")
+
+    if not decision:
+        return [TextContent(type="text", text="Error: decision is required")]
+    if not rationale:
+        return [TextContent(type="text", text="Error: rationale is required")]
+    if reversibility not in ("reversible", "hard", "irreversible"):
+        return [
+            TextContent(
+                type="text",
+                text="Error: reversibility must be one of reversible|hard|irreversible",
+            )
+        ]
+
+    confidence = args.get("confidence")
+    if confidence is not None:
+        try:
+            confidence = float(confidence)
+        except (TypeError, ValueError):
+            return [TextContent(type="text", text="Error: confidence must be a number")]
+        if not (0.0 <= confidence <= 1.0):
+            return [TextContent(type="text", text="Error: confidence must be in [0.0, 1.0]")]
+
+    actor = args.get("actor") or "skill:unknown"
+    project = args.get("project")
+
+    client = server._get_client()
+
+    resolved_memories, unresolved_memories = _resolve_memory_refs(
+        client, args.get("memories_used") or [], project
+    )
+
+    payload = {
+        "decision": decision,
+        "rationale": rationale,
+        "memories_used": resolved_memories,
+        "outcomes_referenced": args.get("outcomes_referenced") or [],
+        "alternatives_considered": args.get("alternatives_considered") or [],
+        "reversibility": reversibility,
+    }
+    if unresolved_memories:
+        payload["memories_used_unresolved"] = unresolved_memories
+    if confidence is not None:
+        payload["confidence"] = confidence
+    if project:
+        payload["project"] = project
+
+    try:
+        result = (
+            client.table("episodes")
+            .insert(
+                {
+                    "actor": actor,
+                    "kind": "decision_made",
+                    "payload": payload,
+                }
+            )
+            .execute()
+        )
+    except Exception as exc:
+        return [TextContent(type="text", text=f"Error recording decision: {exc}")]
+
+    if not result.data:
+        return [TextContent(type="text", text="Failed to record decision.")]
+
+    eid = result.data[0].get("id", "?")
+    msg = f"Decision recorded: episode {eid}"
+    if unresolved_memories:
+        msg += (
+            f" (warning: {len(unresolved_memories)} memory name(s) did not "
+            f"resolve to UUIDs: {unresolved_memories} — check spelling or "
+            "pass memory UUID from recall)"
+        )
+    return [TextContent(type="text", text=msg)]
