+++ mcp-memory/tools_schema.py
@@ -1,0 +1,726 @@
+"""MCP tool schemas (#360 split).
+
+Pure data: returns the list of `Tool` objects registered by the
+memory MCP server. Kept separate from server.py so the entry stays
+thin. Shape is identical to the original @server.list_tools() body —
+no schema changes.
+"""
+
+from __future__ import annotations
+
+from mcp.types import Tool
+
+# Validation tuples used by enum lists below — duplicated from server.py
+# so this module has no runtime dependency on it (avoids circular load).
+VALID_TYPES = ("user", "project", "decision", "feedback", "reference")
+VALID_GOAL_PRIORITIES = ("P0", "P1", "P2")
+VALID_GOAL_STATUSES = ("active", "achieved", "paused", "abandoned")
+
+
+def tool_definitions() -> list[Tool]:
+    return [
+        # ---- Goal tools ----
+        Tool(
+            name="goal_set",
+            description=(
+                "Create or update a goal (upsert by slug). "
+                "Goals are strategic objectives that guide Jarvis's priorities and decisions."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "slug": {
+                        "type": "string",
+                        "description": "Unique identifier (e.g. 'redrobot-demo', 'jarvis-goals-system')",
+                    },
+                    "title": {"type": "string", "description": "Human-readable goal title"},
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Project scope (e.g. 'redrobot', 'jarvis'). null = cross-project.",
+                    },
+                    "direction": {
+                        "type": ["string", "null"],
+                        "description": "Strategic direction this goal belongs to",
+                    },
+                    "priority": {
+                        "type": "string",
+                        "enum": list(VALID_GOAL_PRIORITIES),
+                        "description": "P0 = critical, P1 = important, P2 = nice to have",
+                    },
+                    "status": {
+                        "type": "string",
+                        "enum": list(VALID_GOAL_STATUSES),
+                    },
+                    "why": {"type": "string", "description": "Motivation — why this goal matters"},
+                    "success_criteria": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "List of success criteria",
+                    },
+                    "deadline": {
+                        "type": ["string", "null"],
+                        "description": "Deadline date (YYYY-MM-DD)",
+                    },
+                    "progress": {
+                        "type": "array",
+                        "items": {
+                            "type": "object",
+                            "properties": {
+                                "item": {"type": "string"},
+                                "done": {"type": "boolean"},
+                            },
+                        },
+                        "description": "Progress milestones",
+                    },
+                    "progress_pct": {
+                        "type": "integer",
+                        "description": "Overall progress percentage (0-100)",
+                    },
+                    "risks": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "Known risks",
+                    },
+                    "owner_focus": {
+                        "type": "string",
+                        "description": "What the owner is working on",
+                    },
+                    "jarvis_focus": {"type": "string", "description": "What Jarvis should handle"},
+                    "parent_id": {
+                        "type": ["string", "null"],
+                        "description": "Parent goal UUID (for sub-goals)",
+                    },
+                },
+                "required": ["slug", "title"],
+            },
+        ),
+        Tool(
+            name="goal_list",
+            description=(
+                "List goals with optional filters. "
+                "Use at session start to load active goals as strategic context."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "status": {
+                        "type": "string",
+                        "enum": list(VALID_GOAL_STATUSES),
+                        "description": "Filter by status (default: all)",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Filter by project",
+                    },
+                    "priority": {
+                        "type": "string",
+                        "enum": list(VALID_GOAL_PRIORITIES),
+                        "description": "Filter by priority",
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="goal_get",
+            description="Get a specific goal by slug with full details.",
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "slug": {"type": "string", "description": "Goal slug"},
+                },
+                "required": ["slug"],
+            },
+        ),
+        Tool(
+            name="goal_update",
+            description=(
+                "Partial update of a goal. Only provided fields are updated. "
+                "Use to update progress, status, focus, risks, etc."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "slug": {"type": "string", "description": "Goal slug to update"},
+                    "title": {"type": "string"},
+                    "priority": {"type": "string", "enum": list(VALID_GOAL_PRIORITIES)},
+                    "status": {"type": "string", "enum": list(VALID_GOAL_STATUSES)},
+                    "why": {"type": "string"},
+                    "success_criteria": {"type": "array", "items": {"type": "string"}},
+                    "deadline": {"type": ["string", "null"]},
+                    "progress": {
+                        "type": "array",
+                        "items": {
+                            "type": "object",
+                            "properties": {
+                                "item": {"type": "string"},
+                                "done": {"type": "boolean"},
+                            },
+                        },
+                    },
+                    "progress_pct": {"type": "integer"},
+                    "risks": {"type": "array", "items": {"type": "string"}},
+                    "owner_focus": {"type": "string"},
+                    "jarvis_focus": {"type": "string"},
+                    "outcome": {"type": "string", "description": "What happened (for closing)"},
+                    "lessons": {"type": "string", "description": "What was learned (for closing)"},
+                },
+                "required": ["slug"],
+            },
+        ),
+        # ---- Memory tools ----
+        Tool(
+            name="memory_store",
+            description=(
+                "Save or update a memory. Upserts by (project, name). "
+                "Use for: decisions, user preferences, project context, feedback, references. "
+                "Set project=null for cross-project memories."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "type": {
+                        "type": "string",
+                        "enum": list(VALID_TYPES),
+                        "description": "Memory category",
+                    },
+                    "name": {
+                        "type": "string",
+                        "description": "Unique name within project scope (e.g. 'architecture_split', 'user_work_style')",
+                    },
+                    "content": {
+                        "type": "string",
+                        "description": "Full memory content. Be specific — this is what future sessions will read.",
+                    },
+                    "description": {
+                        "type": "string",
+                        "description": "One-line summary for quick relevance matching.",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Project scope. null = global/cross-project. 'jarvis' = this project.",
+                    },
+                    "tags": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "Tags for filtering (e.g. ['architecture', 'decision'])",
+                    },
+                    "source_provenance": {
+                        "type": "string",
+                        "description": (
+                            "Where this memory came from. Required as of Phase 2c — "
+                            "JTMS attribution, we can't revise what we can't attribute. "
+                            "Use a namespaced form: 'session:<id>', 'skill:<name>', "
+                            "'hook:<name>', 'user:explicit', 'episode:<episode_id>' "
+                            "(Phase 4), or a URL/tool-name when external."
+                        ),
+                    },
+                },
+                "required": ["type", "name", "content", "source_provenance"],
+            },
+        ),
+        Tool(
+            name="memory_recall",
+            description=(
+                "Search memories by keyword or semantic meaning. "
+                "Uses vector similarity search when available, falls back to keyword matching. "
+                "Use at the START of a session to load relevant context, "
+                "or when the user references something discussed before."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "query": {
+                        "type": "string",
+                        "description": "Search query — natural language or keywords",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Filter by project. null = search all projects.",
+                    },
+                    "type": {
+                        "type": "string",
+                        "enum": list(VALID_TYPES),
+                        "description": "Filter by memory type.",
+                    },
+                    "limit": {
+                        "type": "integer",
+                        "description": "Max results (default 10)",
+                        "default": 10,
+                    },
+                    "include_links": {
+                        "type": "boolean",
+                        "description": "Include 1-hop linked memories in results",
+                        "default": False,
+                    },
+                    "show_history": {
+                        "type": "boolean",
+                        "description": (
+                            "Include superseded/expired memories. Default false "
+                            "(live memory only). Set true for audit/debug to see "
+                            "what beliefs were once held."
+                        ),
+                        "default": False,
+                    },
+                    "brief": {
+                        "type": "boolean",
+                        "description": (
+                            "When true, omit full content — return only name, "
+                            "type, project, tags, description, and score. Use to "
+                            "preview what's relevant before committing prompt "
+                            "budget; call memory_get for full content on hits."
+                        ),
+                        "default": False,
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="memory_get",
+            description="Get a specific memory by exact name and project.",
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "name": {
+                        "type": "string",
+                        "description": "Exact memory name",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Project scope. null = global.",
+                    },
+                },
+                "required": ["name"],
+            },
+        ),
+        Tool(
+            name="memory_list",
+            description=(
+                "List all memories, optionally filtered by project and/or type. "
+                "Returns name + description (not full content) for quick overview."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Filter by project. Omit to list all.",
+                    },
+                    "type": {
+                        "type": "string",
+                        "enum": list(VALID_TYPES),
+                        "description": "Filter by type.",
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="memory_delete",
+            description="Soft-delete a memory by name. Recoverable for 30 days via memory_restore.",
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "name": {
+                        "type": "string",
+                        "description": "Memory name to delete",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Project scope. null = global.",
+                    },
+                },
+                "required": ["name"],
+            },
+        ),
+        Tool(
+            name="memory_restore",
+            description="Restore a soft-deleted memory within the 30-day retention window.",
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "name": {
+                        "type": "string",
+                        "description": "Memory name to restore",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Project scope. null = global.",
+                    },
+                },
+                "required": ["name"],
+            },
+        ),
+        # ---- Event tools ----
+        Tool(
+            name="events_list",
+            description=(
+                "List events from the event queue. By default returns unprocessed events "
+                "sorted by severity. GitHub Actions write events here; the orchestrator reads them."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "repo": {
+                        "type": "string",
+                        "description": "Filter by repo (e.g. 'Osasuwu/jarvis')",
+                    },
+                    "event_type": {
+                        "type": "string",
+                        "description": "Filter by event type (e.g. 'ci_failure', 'pr_approved')",
+                    },
+                    "severity": {
+                        "type": "string",
+                        "enum": ["critical", "high", "medium", "low", "info"],
+                        "description": "Filter by minimum severity",
+                    },
+                    "include_processed": {
+                        "type": "boolean",
+                        "description": "Include already-processed events (default: false)",
+                    },
+                    "limit": {
+                        "type": "integer",
+                        "description": "Max results (default 20)",
+                        "default": 20,
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="events_mark_processed",
+            description=(
+                "Mark one or more events as processed. "
+                "Call after the orchestrator has handled an event."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "event_ids": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "List of event UUIDs to mark as processed",
+                    },
+                    "processed_by": {
+                        "type": "string",
+                        "description": "Who processed it (e.g. 'autonomous-loop', 'risk-radar', 'manual')",
+                    },
+                    "action_taken": {
+                        "type": "string",
+                        "description": "What was done in response",
+                    },
+                },
+                "required": ["event_ids", "processed_by"],
+            },
+        ),
+        # ---- Outcome tracking tools (Pillar 3) ----
+        Tool(
+            name="outcome_record",
+            description=(
+                "Record a task outcome for tracking and learning. "
+                "Call after completing a delegation, fix, research, or autonomous action."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "task_type": {
+                        "type": "string",
+                        "enum": ["delegation", "research", "fix", "review", "autonomous"],
+                        "description": "Type of action performed.",
+                    },
+                    "task_description": {
+                        "type": "string",
+                        "description": "What was done (concise).",
+                    },
+                    "outcome_status": {
+                        "type": "string",
+                        "enum": ["pending", "success", "partial", "failure", "unknown"],
+                        "description": "Outcome: success/partial/failure/pending/unknown.",
+                    },
+                    "outcome_summary": {
+                        "type": "string",
+                        "description": "What actually happened.",
+                    },
+                    "goal_slug": {
+                        "type": ["string", "null"],
+                        "description": "Related goal slug.",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Project scope.",
+                    },
+                    "issue_url": {"type": "string", "description": "GitHub issue URL."},
+                    "pr_url": {"type": "string", "description": "GitHub PR URL."},
+                    "tests_passed": {"type": "boolean"},
+                    "pr_merged": {"type": "boolean"},
+                    "quality_score": {
+                        "type": "integer",
+                        "description": "Quality 0-100.",
+                    },
+                    "lessons": {"type": "string", "description": "What was learned."},
+                    "pattern_tags": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "Pattern tags for learning.",
+                    },
+                    "memory_id": {
+                        "type": ["string", "null"],
+                        "description": (
+                            "UUID of the primary memory that informed this outcome. "
+                            "Links the outcome into memory_calibration so confidence "
+                            "and lessons can be attributed back to the reasoning basis."
+                        ),
+                    },
+                },
+                "required": ["task_type", "task_description", "outcome_status"],
+            },
+        ),
+        Tool(
+            name="outcome_update",
+            description=(
+                "Update a task outcome after verification. Use to flip status from pending "
+                "to success/failure, record verified_at, pr_merged, lessons, etc."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "id": {"type": "string", "description": "Outcome UUID to update."},
+                    "outcome_status": {
+                        "type": "string",
+                        "enum": ["pending", "success", "partial", "failure", "unknown"],
+                    },
+                    "outcome_summary": {"type": "string"},
+                    "pr_merged": {"type": "boolean"},
+                    "tests_passed": {"type": "boolean"},
+                    "quality_score": {"type": "integer", "description": "0-100."},
+                    "lessons": {"type": "string"},
+                    "pattern_tags": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                    },
+                    "verified_at": {
+                        "type": "string",
+                        "description": "ISO timestamp. Defaults to now() if omitted when status changes.",
+                    },
+                    "memory_id": {
+                        "type": ["string", "null"],
+                        "description": (
+                            "UUID of the primary memory that informed this outcome. "
+                            "Pass during verification to retro-link outcomes whose "
+                            "basis became clear only after the decision played out."
+                        ),
+                    },
+                },
+                "required": ["id"],
+            },
+        ),
+        Tool(
+            name="outcome_list",
+            description=(
+                "List recent task outcomes, optionally filtered by project, goal, status, or pattern_tags. "
+                "Use to review what worked and what didn't."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Filter by project.",
+                    },
+                    "goal_slug": {
+                        "type": ["string", "null"],
+                        "description": "Filter by goal slug.",
+                    },
+                    "outcome_status": {
+                        "type": "string",
+                        "enum": ["pending", "success", "partial", "failure", "unknown"],
+                    },
+                    "pattern_tag": {
+                        "type": "string",
+                        "description": "Filter by pattern tag (outcomes containing this tag).",
+                    },
+                    "limit": {
+                        "type": "integer",
+                        "description": "Max results (default 20).",
+                        "default": 20,
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="memory_calibration_summary",
+            description=(
+                "Confidence calibration summary: Brier score of predicted vs actual outcomes, "
+                "bucketed by memory type. Reveals systemic over- or under-confidence. "
+                "Used by /reflect and /self-improve (#251)."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Optional project filter. null/omitted = global.",
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="record_decision",
+            description=(
+                "Record a decision made by the agent as a 'decision_made' episode. "
+                "Captures decision text, rationale, memory/outcome IDs that informed it, "
+                "predicted confidence (0.0-1.0), alternatives, and reversibility. "
+                "Feeds the reasoning-trace for later /reflect analysis (#252)."
+            ),
+            inputSchema={
+                "type": "object",
+                "required": ["decision", "rationale", "reversibility"],
+                "properties": {
+                    "decision": {
+                        "type": "string",
+                        "description": "Short statement of what was decided.",
+                    },
+                    "rationale": {
+                        "type": "string",
+                        "description": "One-paragraph why — the basis for the choice.",
+                    },
+                    "memories_used": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "Memory IDs that informed this decision (from recall).",
+                    },
+                    "outcomes_referenced": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "task_outcomes IDs that informed this decision.",
+                    },
+                    "confidence": {
+                        "type": "number",
+                        "minimum": 0.0,
+                        "maximum": 1.0,
+                        "description": "Predicted confidence the decision is correct (0.0-1.0).",
+                    },
+                    "alternatives_considered": {
+                        "type": "array",
+                        "items": {"type": "string"},
+                        "description": "Options rejected and briefly why.",
+                    },
+                    "reversibility": {
+                        "type": "string",
+                        "enum": ["reversible", "hard", "irreversible"],
+                        "description": "How easily this decision can be undone.",
+                    },
+                    "actor": {
+                        "type": ["string", "null"],
+                        "description": "Source of the decision (e.g. 'skill:delegate', 'session:<id>'). Defaults to 'skill:unknown'.",
+                    },
+                    "project": {
+                        "type": ["string", "null"],
+                        "description": "Optional project scope for the decision payload.",
+                    },
+                },
+            },
+        ),
+        # ---- Graph tools ----
+        Tool(
+            name="memory_graph",
+            description=(
+                "Explore the memory link graph. "
+                "Modes: 'overview' (stats, top connected, orphans), "
+                "'links' (all connections for a specific memory by name), "
+                "'clusters' (groups of tightly connected memories for consolidation)."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "mode": {
+                        "type": "string",
+                        "enum": ["overview", "links", "clusters"],
+                        "description": (
+                            "overview = link stats + top connected + orphans. "
+                            "links = all connections for a memory (requires 'name'). "
+                            "clusters = tightly connected groups."
+                        ),
+                    },
+                    "name": {
+                        "type": "string",
+                        "description": "Memory name (required for 'links' mode).",
+                    },
+                },
+                "required": ["mode"],
+            },
+        ),
+        # ---- Credential registry tools (Pillar 9) ----
+        Tool(
+            name="credential_list",
+            description=(
+                "List registered credentials (metadata only — never returns secret values). "
+                "Shows service name, env var name, storage location, expiry, rotation notes."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "scope": {
+                        "type": ["string", "null"],
+                        "description": "Filter by scope (e.g. 'jarvis'). null = all.",
+                    },
+                },
+            },
+        ),
+        Tool(
+            name="credential_add",
+            description=(
+                "Register a credential in the metadata-only registry. "
+                "Stores service name, env var name, where it's kept, and rotation info. "
+                "NEVER pass actual secret values — the table rejects them."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "service": {
+                        "type": "string",
+                        "description": "Service name (e.g. 'Supabase', 'GitHub')",
+                    },
+                    "env_var": {
+                        "type": "string",
+                        "description": "Env variable NAME, not value (e.g. 'SUPABASE_KEY')",
+                    },
+                    "stored_in": {
+                        "type": "string",
+                        "description": "Where the value lives (e.g. '.env', 'GitHub Actions', 'system env')",
+                        "default": ".env",
+                    },
+                    "scope": {
+                        "type": "string",
+                        "description": "Project scope (e.g. 'jarvis')",
+                        "default": "jarvis",
+                    },
+                    "expires_at": {
+                        "type": ["string", "null"],
+                        "description": "Expiry date ISO format (null = no expiry)",
+                    },
+                    "rotation_notes": {
+                        "type": ["string", "null"],
+                        "description": "How to rotate (e.g. 'Anthropic Console → API Keys')",
+                    },
+                    "notes": {"type": ["string", "null"], "description": "Additional notes"},
+                },
+                "required": ["service", "env_var"],
+            },
+        ),
+        Tool(
+            name="credential_check_expiry",
+            description=(
+                "Check for credentials expiring within N days. "
+                "Returns list with service, env var, expiry date, and rotation notes. "
+                "Use in morning-brief for proactive alerts."
+            ),
+            inputSchema={
+                "type": "object",
+                "properties": {
+                    "days_ahead": {
+                        "type": "integer",
+                        "description": "Alert window in days (default 30)",
+                        "default": 30,
+                    },
+                },
+            },
+        ),
+    ]
